rename: ReadWriteSpliting → ReadWriteSplitting
