+++ src/BuildingBlocks/Ddd/Domain/Masa.BuildingBlocks.Ddd.Domain/Masa.BuildingBlocks.Ddd.Domain.csproj
@@ -10,7 +10,7 @@
     <ProjectReference Include="..\..\..\Data\Masa.BuildingBlocks.Data.Contracts\Masa.BuildingBlocks.Data.Contracts.csproj" />
     <ProjectReference Include="..\..\..\Data\Masa.BuildingBlocks.Data.UoW\Masa.BuildingBlocks.Data.UoW.csproj" />
     <ProjectReference Include="..\..\..\Dispatcher\Masa.BuildingBlocks.Dispatcher.IntegrationEvents\Masa.BuildingBlocks.Dispatcher.IntegrationEvents.csproj" />
-    <ProjectReference Include="..\..\..\ReadWriteSpliting\Masa.BuildingBlocks.ReadWriteSpliting.Cqrs\Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.csproj" />
+    <ProjectReference Include="..\..\..\ReadWriteSplitting\Masa.BuildingBlocks.ReadWriteSplitting.Cqrs\Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.csproj" />
   </ItemGroup>
