+++ src/BuildingBlocks/Ddd/Domain/Masa.BuildingBlocks.Ddd.Domain/_Imports.cs
@@ -8,8 +8,8 @@ global using Masa.BuildingBlocks.Ddd.Domain.Entities.Auditing;
 global using Masa.BuildingBlocks.Ddd.Domain.Events;
 global using Masa.BuildingBlocks.Dispatcher.Events;
 global using Masa.BuildingBlocks.Dispatcher.IntegrationEvents;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Commands;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Queries;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Commands;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Queries;
 global using System.Linq.Expressions;
 global using System.Reflection;
 global using System.Text.Json.Serialization;
