+++ src/BuildingBlocks/ReadWriteSplitting/Masa.BuildingBlocks.ReadWriteSplitting.Cqrs/Commands/ICommand.cs
@@ -1,7 +1,7 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Commands;
+namespace Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Commands;
 
 public interface ICommand : IEvent, ITransaction
 {
