+++ src/BuildingBlocks/ReadWriteSplitting/Masa.BuildingBlocks.ReadWriteSplitting.Cqrs/Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.csproj
