+++ src/BuildingBlocks/ReadWriteSplitting/Masa.BuildingBlocks.ReadWriteSplitting.Cqrs/Queries/IQueryHandler.cs
@@ -1,7 +1,7 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Queries;
+namespace Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Queries;
 
 public interface IQueryHandler<TCommand, TResult> : IEventHandler<TCommand>
     where TCommand : IQuery<TResult>
