+++ src/BuildingBlocks/ReadWriteSplitting/Masa.BuildingBlocks.ReadWriteSplitting.EventSourcing/Masa.BuildingBlocks.ReadWriteSpliting.EventSourcing.csproj
