--- src/Contrib/ReadWriteSpliting/Tests/Masa.Contrib.ReadWriteSpliting.Cqrs.Tests/_Imports.cs
@@ -1,13 +1,0 @@
-// Copyright (c) MASA Stack All rights reserved.
-// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
-
-global using Masa.BuildingBlocks.Dispatcher.Events;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Commands;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Queries;
-global using Masa.Contrib.Dispatcher.Events;
-global using Masa.Contrib.ReadWriteSpliting.Cqrs.Commands;
-global using Masa.Contrib.ReadWriteSpliting.Cqrs.Queries;
-global using Masa.Contrib.ReadWriteSpliting.Cqrs.Tests.Commands;
-global using Masa.Contrib.ReadWriteSpliting.Cqrs.Tests.Queries;
-global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
