+++ src/Contrib/ReadWriteSplitting/Masa.Contrib.ReadWriteSplitting.Cqrs/Commands/CommandHandler.cs
@@ -1,7 +1,7 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Contrib.ReadWriteSpliting.Cqrs.Commands;
+namespace Masa.Contrib.ReadWriteSplitting.Cqrs.Commands;
 
 public abstract class CommandHandler<TCommand> : ICommandHandler<TCommand>, ISagaEventHandler<TCommand>
     where TCommand : ICommand
