+++ src/Contrib/ReadWriteSplitting/Masa.Contrib.ReadWriteSplitting.Cqrs/Masa.Contrib.ReadWriteSplitting.Cqrs.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <RootNamespace>Masa.Contrib.ReadWriteSplitting.Cqrs</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\BuildingBlocks\ReadWriteSpliting\Masa.BuildingBlocks.ReadWriteSpliting.Cqrs\Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.csproj" />
+    <ProjectReference Include="..\..\..\BuildingBlocks\ReadWriteSplitting\Masa.BuildingBlocks.ReadWriteSplitting.Cqrs\Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.csproj" />
   </ItemGroup>
 
 </Project>
