+++ src/Contrib/ReadWriteSplitting/Masa.Contrib.ReadWriteSplitting.Cqrs/README.md
@@ -1,19 +1,19 @@
-中 | [EN](README.md)
+[中](README.zh-CN.md) | EN
 
 ## Cqrs
 
-用例：
+Example：
 
-1. 定义了Command与Query基类
-2. 支持Saga模式的Handler，并提供CommandHandler基础实现
+1. Define Command and Query base classes
+2. Support the Handler of Saga mode, and provide the basic implementation of CommandHandler
 
    ```C#
-   Install-Package Masa.Contrib.ReadWriteSpliting.Cqrs
+   Install-Package Masa.Contrib.ReadWriteSplitting.Cqrs
    ```
 
 ##### Query：
 
-1. 定义Query
+1. Define Query
 
    ```C#
    public class CatalogItemQuery : Query<List<CatalogItem>>
@@ -24,15 +24,14 @@
    }
    ```
 
-2. 定义QueryHandler
+2. Define QueryHandler
 
    ```C#
    public class CatalogQueryHandler : QueryHandler<CatalogItemQuery, List<CatalogItem>>
    {
        private readonly ICatalogItemRepository _catalogItemRepository;
 
-       public CatalogQueryHandler(ICatalogItemRepository catalogItemRepository)
-           => _catalogItemRepository = catalogItemRepository;
+       public CatalogQueryHandler(ICatalogItemRepository catalogItemRepository) => _catalogItemRepository = catalogItemRepository;
 
        public async Task HandleAsync(CatalogItemQuery query)
        {
@@ -41,18 +40,18 @@
    }
    ```
 
-3. 发送Query
+3. Send Query
 
    ```c#
-   IEventBus eventBus;//通过DI得到IEventBus
+   IEventBus eventBus;//Get IEventBus through DI
    await eventBus.PublishAsync(new CatalogItemQuery() { Name = "Rolex" });
    ```
 
-   > 提示：Query后的泛型与Result的返回类型保持一致，需要再Handler中为Result赋值，以便调用方得到结果
+   > Tip: The generic type after Query is consistent with the return type of Result. You need to assign a value to Result in Handler so that the caller can get the result.
 
 ##### Command
 
-1. 定义 Command
+1. Define Command
 
    ```c#
    public class CreateCatalogItemCommand : Command
@@ -63,14 +62,14 @@
    }
    ```
 
-2. 添加 CommandHandler
+2. Add CommandHandler
 
    ```c#
    public class CatalogCommandHandler : CommandHandler<CreateCatalogItemCommand>
    {
        private readonly ICatalogItemRepository _catalogItemRepository;
 
-       public CatalogCommandHandler(ICatalogItemRepository catalogItemRepository) => _catalogItemRepository =    catalogItemRepository;
+       public CatalogCommandHandler(ICatalogItemRepository catalogItemRepository) => _catalogItemRepository = catalogItemRepository;
 
        public async Task HandleAsync(CreateCatalogItemCommand command)
        {
@@ -79,9 +78,9 @@
    }
    ```
 
-3. 发送 Command
+3. Send Command
 
    ```C#
-   IEventBus eventBus;//通过DI得到IEventBus
+   IEventBus eventBus;//Get IEventBus through DI
    await eventBus.PublishAsync(new CreateCatalogItemCommand());
    ```
