+++ src/Contrib/ReadWriteSplitting/Masa.Contrib.ReadWriteSplitting.Cqrs/_Imports.cs
@@ -2,5 +2,5 @@
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
 global using Masa.BuildingBlocks.Dispatcher.Events;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Commands;
-global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Queries;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Commands;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Queries;
