+++ src/Contrib/ReadWriteSplitting/Tests/Masa.Contrib.ReadWriteSplitting.Cqrs.Tests/ProductionQueryHandler.cs
@@ -1,7 +1,7 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Contrib.ReadWriteSpliting.Cqrs.Tests;
+namespace Masa.Contrib.ReadWriteSplitting.Cqrs.Tests;
 
 public class ProductionQueryHandler : QueryHandler<ProductionItemQuery, string>
 {
