+++ src/Contrib/ReadWriteSplitting/Tests/Masa.Contrib.ReadWriteSplitting.Cqrs.Tests/_Imports.cs
@@ -1,0 +1,13 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+global using Masa.BuildingBlocks.Dispatcher.Events;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Commands;
+global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Queries;
+global using Masa.Contrib.Dispatcher.Events;
+global using Masa.Contrib.ReadWriteSplitting.Cqrs.Commands;
+global using Masa.Contrib.ReadWriteSplitting.Cqrs.Queries;
+global using Masa.Contrib.ReadWriteSplitting.Cqrs.Tests.Commands;
+global using Masa.Contrib.ReadWriteSplitting.Cqrs.Tests.Queries;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
